Fix CodeQL issues and add value clamping for range widgets
- LinuxProgressBar: Clamp value when Minimum/Maximum changes
- LinuxScrollBar: Clamp value in UpdateAdjustment(), add readonly to _adjustment
- LinuxSash: Initialize _requestedBounds in constructor
- LinuxScale: Cast to double before multiplication to prevent overflow
- LinuxList: Combine nested if statements
- BrowserTests: Use 'using' for proper disposal, polling instead of blocking
- Add CodeQL config to suppress expected unmanaged code warnings

+++ src/SWTSharp/Platform/Linux/LinuxScale.cs
@@ -135,7 +135,7 @@ internal class LinuxScale : LinuxWidget, IPlatformScale
             if (value > 0 && _increment != value)
             {
                 _increment = value;
-                gtk_range_set_increments(_scale, _increment, _increment * 10);
+                gtk_range_set_increments(_scale, _increment, (double)_increment * 10);
                 UpdateTickMarks();
             }
         }
